backend/requirements.txt
fastapi
mangum
pydantic
# boto3 is provided by Lambda runtime
uvicorn
python-jose[cryptography]
requests
